Add files via upload
In this project I implemented DNN and CNN to classify images of hands in poses that correspond to letters in American Sign Language. The dataset is contained in the  attached zip file, along with some images and a text file describing the dataset. It is similar in many ways to other MNIST datasets.